Fix unsigned short overflow
Replaced unsigned short with unsigned int to hold indices.

+++ include/Lattice.hpp
@@ -31,8 +31,8 @@ public:
   template<std::size_t S>
   bool input(std::array<float, S> in);
   std::vector<Node> const& neurons() const;
-  std::vector<unsigned short> lineIndices() const;
-  std::vector<unsigned short> triangleIndices() const;
+  std::vector<unsigned int> lineIndices() const;
+  std::vector<unsigned int> triangleIndices() const;
 };
 
 template<std::size_t S>
